More warning fixes and enable 68020 for debugging

+++ uae4all_gp2x_0.7.2a/src/m68k/fame/famec_opcodes.h
@@ -3,7 +3,7 @@
 
 #define EXECUTE_EXCEPTION(EX,CYCLES)   \
 {                                      \
-u32 oldPC=(u32)GET_PC;                   \
+u32 oldPC=GET_PC;                   \
 SET_PC(oldPC-2)                     \
 execute_exception(EX);              \
 RET(CYCLES)                         \
